build: upgrade cypress and split tests
