+++ superset-frontend/cypress-base/package.json
@@ -10,10 +10,10 @@
   "license": "Apache-2.0",
   "dependencies": {
     "shortid": "^2.2.15",
-    "@cypress/code-coverage": "^3.7.6"
+    "@cypress/code-coverage": "^3.8.1"
   },
   "devDependencies": {
-    "cypress": "4.7.0",
+    "cypress": "4.9.0",
     "eslint-plugin-cypress": "^2.11.1"
   }
 }
